+++ src/components/LandingPage/AIPredictions/AIPredictions.css
