+++ src/components/LandingPage/AIPredictions/AIPredictions.tsx
