+++ src/pages/LandingPage.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import Navbar from '../components/Navbar/Navbar';
-import Hero from '../components/Hero/Hero';
-import WhyLionTrade from '../components/WhyLionTrade/WhyLionTrade';
+import Navbar from '../components/LandingPage/Navbar/Navbar';
+import Hero from '../components/LandingPage/Hero/Hero';
+import WhyLionTrade from '../components/LandingPage/WhyLionTrade/WhyLionTrade';
 import './LandingPage.css';
-import LiveMarketInsights from "../components/LiveMarketInsights/LiveMarketInsights.tsx";
-import AIPredictions from "../components/AIPredictions/AIPredictions.tsx";
-import PricingPlans from "../components/PricingPlans/PricingPlans.tsx";
-import Footer from "../components/Footer/Footer.tsx";
-import JoinCTA from "../components/JoinCTA/JoinCTA.tsx";
+import LiveMarketInsights from "../components/LandingPage/LiveMarketInsights/LiveMarketInsights.tsx";
+import AIPredictions from "../components/LandingPage/AIPredictions/AIPredictions.tsx";
+import PricingPlans from "../components/LandingPage/PricingPlans/PricingPlans.tsx";
+import Footer from "../components/LandingPage/Footer/Footer.tsx";
+import JoinCTA from "../components/LandingPage/JoinCTA/JoinCTA.tsx";
 
 const LandingPage: React.FC = () => {
     return (
